Moved the HeadingParagraphStyle to the model package
