+++ richeditor-compose/src/commonMain/kotlin/com/mohamedrejeb/richeditor/parser/utils/ElementsParagraphStyle.kt
@@ -1,5 +1,7 @@
 package com.mohamedrejeb.richeditor.parser.utils
 
+import com.mohamedrejeb.richeditor.model.HeadingParagraphStyle
+
 internal val H1ParagraphStyle = HeadingParagraphStyle.H1.getParagraphStyle()
 internal val H2ParagraphStyle = HeadingParagraphStyle.H2.getParagraphStyle()
 internal val H3ParagraphStyle = HeadingParagraphStyle.H3.getParagraphStyle()
